src/requirements.txt
langchain
python-dotenv
supabase
aiohttp
ccxt
fastapi
tqdm
appwrite
apscheduler
apscheduler